Independent GoThroughPosition ASV task
Working version.

In this branch, the implementation is simplified only for ASV to not
affect other platofroms that relly on the common implementation
It allows to have simplified observation buffers and faster training.
The Core and common GoThroughPosition tasks have separate observations
buffer segments that are designed to handle multiple tasks. Here, the
GoThroughPosition tasks handles the buffer update instead of the Core to
reduce the dimentionality of the observations space.