Fix submerged volume computation and USD mass
The sumerged volume computation was based on box_width and box_length
instead of the waterplane area and draught. The submerged volume was
also potentially being computed using wrong units (centimeters). The
implemented fix uses aproximate hull dimensions for the hulls and also
updates the mass of the heron in the USD file that was being
autocomputed. By changing this values, the rotation of the ASV changed
sgnificantly becoming much more realistic.