Fix Mass and CoM randomization
The randomized mass and com was not being applied to the platform.

To keep the randomization code agnostic to the simulation environment
(Isaac Sim) the setting of values was implemented in the ASV_Virtual.py
and not in the MassDistributionDisturbances class.

Minimal adjustments were made to the tasks/common_3DoF/disturbances.py
file to prevent inconsistency with other platforms.