Add moment of inertia randomization
Follow same structre that was already implemented for 'mass' randomization
which is actually a mass distribution type of randomization that
includes the randomization of the mass itself, the center of mass, and
the moment of inertia.

This implementation randomizes only one dimention of the moment of
inertia around the Z axis.