Lower the linear rotation drag.
Previous value was optimized to match the maximum rotational velocity
when the boat is moving in a circle. For lower linear velocities the
value is too high and prevents the boat from rotating in simulation.
The new value is a better compromise between the two. Robustness should
be achieved by randomizing this value during trainning